Added support for both 'IP' and 'IPAdress' directives in [Network] sections. This was done to accomodate my own tendency to use 'IP' instead of 'IPAdress'. In case both 'IP' and 'IPAdress' are defined in the section then the 'IPAdress will take precedent.
Long-term the solution should be to treat both directives as aliases for
each other and to develop a more robust way of handling Aliases/synonymous
directives. In this case we should try to apply each of the IPAddresseses
to the adapter (although this creates an issue determining which netmask
belongs to which address), as this would allow us to deal withmore complex
networking setups.