Added CAFForceInteractive and tweaked Verify-Assertions.
Added the CAFForceInteractive as a special operation that attempts to
start Caffeine in any 'interactive' user session on the machine to
support installers that will not run in non-interactive (command-line
only) sessions. This hinges on the credentials for an active interactive
sessions being present in the configuration/job file.

Also tweaked Verify-Assertions to add backwards compatibility with 'WS2012R2'