Started splitting Caffeine into to separate projects.
The core of Caffeine is an engine that takes a special .ini file and
performs setup according to instructions in that file, however it has
always been limited to the task sequence hardcoded into caffeine.ps1.

The goal is to make Caffeine completely independent of the original task
sequence that it was developed around: CAF ("Collect And Analyze").

CAF was a continuation of "MOCsetup", and is primarily designed around
your typical Microsoft Official Courseware Hyper-V environment. The CAF
TS is however very inefficient for many other types of setups, which may
benefit from specially tailored sequences.