Modularized Caffeine
The intended way to use Caffeine is to install it using the
Install-Caffeine command. This will cause Caffeine to run at
startup as SYSTEM (using a Scheduled Task).

Alternatively, Caffeine can be run manually using the Start-Caffeine
command.

Either way, from this point on any reference to Caffeinate.ps1 is
considered obsolete and should be replaced with calls to
Install-Caffeine or Run-Caffeine as appropriate.